requirements.txt
openai~=1.42.0
python-fasthtml~=0.4.4